chore: move web to src directory
