+++ src/web/convert/breakpoint.ts
@@ -2,7 +2,7 @@ import type { MediaQuery } from 'typestyle/lib/types'
 import { UnistylesState } from '../state'
 import { keyInObject } from '../utils'
 import { isUnistylesMq, parseMq } from '../mq'
-import { Orientation } from '../../src/specs/types'
+import { Orientation } from '../../specs/types'
 
 export const convertBreakpoint = (breakpoint: string): MediaQuery => {
     if (Object.values(Orientation).includes(breakpoint as Orientation)) {
